requirements.txt
flask
gspread
google-auth
python-dotenv
line-bot-sdk